HACK: Modify storage device spinup timeout
Previously the timeout was raised to 20s to work around a quirk on a
certain SSD after power wash / dev mode transition. After discussion
with the drive vendor, the worst-case delay is actually 23s. Though we
have not yet seen delays > 14s, bump the timeout to 25s to accomodate
the worst-case theoretical delay.

BUG=chrome-os-partner:23267
TEST=Compile only.

Change-Id: I2c155dcb07d15cec43ef063302f601711b649a4c
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174957
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>