Support eMMC devices requiring exact OCR setting
Some eMMC devices (KLMCG8GE2A-A001 in particular) do not allow
iterative setting of voltages in the OCR register. In fact JEDEC
84-B45 standard mentions in section 6.4.2

"Dual voltage devices may fail if the voltage range 1.95V to 2.7 V is
used."

This change introduces the ability to preset the voltage to known
values and avoid the need to iterate through different OCR settings.

BUG=chrome-os-partner:22580
TEST=manual

  . with this and other changes the KLMCG8GE2A-A001 eMMC device gets
    successfully initialized on Bayley Bay CRB.

Change-Id: Ie661656ae5e2d987559320e713a33b95870c9325
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175174
Reviewed-by: Hung-Te Lin <hungte@chromium.org>