timer: Halt for timer_hz < 10^6
timer_hz < 10^6 means something is misconfigured and it leads to a div0
exception in timer_us() that can be tricky to debug. Check for this case
and halt. Also, since the CPU timer frequency should not change, only
check it once and store the value for later use.

TEST=Manual on Rambi. Force timer_hz to 200,000 and verify printf + halt
encountered instead of div0 exception.
BUG=None.

Change-Id: If8606646a41784bef46d2c82104608bb81eeae48
Reviewed-on: https://chromium-review.googlesource.com/174541
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>