Copy u-boot implementation of the SDHCI driver.
This is an as is copy of the ToT Chrome OS u-boot tree. This code
obviously is not even attempted to be compiled.

This copy will make examining the porting changes somewhat easier
(albeit not much, as the previously copied u-boot code has been
heavily modified both in structure and in style).

BUG=chrome-os-partner:22580
TEST=none

Change-Id: I03f8ab816cf4afa0505b5c257f148fafd7afc862
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175191
Reviewed-by: Gabe Black <gabeblack@chromium.org>