netboot: automate NETBOOT_DYNAMIC_PARAMS detection and remove Kconfig
This patch gets rid of the CONFIG_NETBOOT_DYNAMIC_PARAMS Kconfig that
decides whether to use the netboot parameters (TFTP server IP and boot
file name) stored in the binary or use the DHCP supplied options. The
new code will just use hardcoded values if they are there, and DHCP
options if they aren't (which is the default if you didn't run the
update_firmware_settings.py tool on the image).

BUG=None
TEST=None

Change-Id: Icf74f541baf81033b11868be934e856d8d0bfc42
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175874
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>