Add SDHCI ADMA support
This change adds support for Advanced DMA (ADMA), a feature available
on contemporary SDHCI controllers. It is not conditionally compiled
in, the capability is detected at run time and used if available.

The implementation is following "SD Host Controller Simplified
Specification, Version 3.00, February 25, 2011".

Controllers supporting ADMA also support automated generation of the
CMD12 (stop multiblock transfer) command. The mmc layer is being
modified not to issue this command when SDHCI controller is used.

BUG=chrome-os-partner:22580
TEST=manual

  - tested on Rambi in depthcharge, initializing the controller and
    printing out an eMMC block

  - tested identical changes on Bayley Bay in u-boot, extensively
    reading and writing long blocks of data (including the complete
    kernel).

Change-Id: I55f34e7dde118bd67ac6f53e46e0ce4775ec76ef
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175987
Reviewed-by: Hung-Te Lin <hungte@chromium.org>