netboot: Add feature to provide dynamic command line option via TFTP
This patch revises the kernel command line generation in netboot. If
additional command line arguments are supplied via the netboot
parameters bundled with the binary, they work just like they always did
(and you can in fact just supply an empty space as "additional args" to
enforce that). Otherwise, netboot will try to load a 'cmdline.<board>'
file from the TFTP server and use that as a command line (including some
simple formatting such as replacing newlines with spaces, so that you
should be able to use your build_image-generated config.txt with minimal
changes). If this file doesn't exist either, it will finally fall back
to just using the built-in default command line (which is also moved
from Kconfig to a static constant because it's really not board
dependent).

BUG=None
TEST=Made sure all three cases work as expected.

Change-Id: I80089466d3036e698f24f43e2bdffee5b5aa2ebd
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/176375
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>