netboot: Fix DHCP errors
This patch fixes a bug in the DHCP stack that prevented our corp DHCP
server from playing nice with netboot. The BOOTP ClientIP fields is
*not* the right place to state which offered IP address we want in a
DHCP Request. Instead you are supposed to use DHCP option 50 (and it's
theoretically acceptable to leave it out completely as well, but in
practice some server implementations seems to not like that).

Also adds DHCP option 61 (Client ID), since other common DHCP clients
seem to send it too and it helps the server keep track of a lease
(making this as familiar as possible for the server seems like the best
idea). Finally, lowered the DHCP request timeout to 3 seconds to avoid
overlong delays with servers that somehow never respond to Discover on
the first try (looking at you, corp...). Even if the network was really
slow, accidential retransmissions shouldn't break DHCP.

BUG=None
TEST=Confirmed it works with both my dnsmasq and the corp DNS server.
Compared tcpdumps with a normal Snow and confirmed the diff is now
smaller (dhcpcd still sends some extra stuff like DNS requesting options
that we probably really don't need).

Change-Id: Ia1570c6f0285d5fa6c6dc5243a366b6b88d3d0c9
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/176487
Reviewed-by: Gabe Black <gabeblack@chromium.org>