nyan: Move the kernel and depthcharge a little farther down in memory.
The kernel was having trouble with the current positions for unknown reasons.
The size of the kernel was at a pretty round number when it started having
problems which lends credibility to the idea of it overwriting something as it
decompressed itself, but I couldn't find specifically where it would do that.

BUG=None
TEST=Built and booted on nyan from the eMMC on cros/chromeos-3.10 which used
to reliably cause panics which looked like memory corruption.
BRANCH=None

Change-Id: Ic68a3f8025052ba18ea21d57b9315c4c866f6297
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/176552
Reviewed-by: Ronald Minnich <rminnich@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>