FIT: Make it possible to override the preferred kernel compat string.
In situations where the same image may need to run on multiple boards which
have only small differences between them, it might be necessary to pick a
different device tree for the kernel depending on the revision of that
particular board.

This change keeps the kconfig variable which has been used up to this point,
but now that only defines the default compat string. A board can override that
setting as necessary as described above.

BUG=None
TEST=Built and booted on old and new nyans.
BRANCH=None

Change-Id: Idf76dbe4f5982253d3724c670c183682bd7635e2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/177591
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>