FIT: Prototype the DeviceTree type in fit.h.
A pointer to a pointer to a DeviceTree structure is a parameter to the
fit_load function, but no definition of that type was being provided directly.
The type had been available by coincidence which is why it would compile.
Instead of including the header which defines the DeviceTree type, lets just
put a prototype for it in fit.h to cut down on the amount of work the compiler
has to do.

BUG=None
TEST=Built for nyan.
BRANCH=None

Change-Id: I90ee34563d77443adc2628b525d314248c26d060
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/177590
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>