nyan: Override the preferred kernel compat string based on board id.
Instead of having to edit the compat string in the default config, we should
take advantage of the fact that we can tell which revision of the board we're
on and pick the right string automatically.

Also, eliminate the now unnecessary NYAN_IN_A_PIXEL kconfig variable.

BUG=None
TEST=Built and booted on both types of nyan.
BRANCH=None

Change-Id: If7526898d5105d39fb1037c7f58103bb238964e0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/177593
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Tom Warren <twarren@nvidia.com>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>