nyan: Reset using GPIO I5 instead of the PMIC.
The PMIC resets the whole machine, and if there's anything in RAM like info
about a kernel panic it can be damaged that way. We should reset using GPIO I5
which resets just the SOC and the TPM, leaving the contents of memory alone.

BUG=None
TEST=Built and booted on nyan. Forced a power off and saw that it turned off
the machine. Forced a reboot and saw that the machine rebooted.
BRANCH=None

Change-Id: Ibcb496ceadc0eeca0bd4483b5050216b1ca79dc2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/177977
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>