tegra: Rename a local variable to be less confusing.
The name of the local variable used to keep track of the APB DMA controller as
it was set up was the same as a member variable of the DMA controller itself.
That was confusing, so this changes the variable to be "controller" instead of
"channels".

BUG=None
TEST=Built and booted on nyan.
BRANCH=None

Change-Id: I9efdc562d2b834e2ffb4ce07c7bc89f444229d75
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/178200
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>