power: Consolidate some common code into the stub functions.
There was some boilerplate code in the power_off and cold_reboot callbacks
in the power drivers which was repeated in each driver and really has no
reason to change. Instead of repeating that code all over, this change moves
it into the plain power_off() and cold_reboot() functions which call into the
registered power driver. This cuts down on duplication and makes the power
drivers a little easier to write.

BUG=None
TEST=Built and booted on nyan, built for falco and pit.
BRANCH=None

Change-Id: I089f3693a7a3160613702c770e1a3110679c3deb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/178201
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>