vboot: Fix VbExDiskGetInfo callback's use of malloc.
A recent patch switched over to a new self checking version of malloc, but in
VbExDiskGetInfo it's possible for malloc to be called with zero bytes, for
that to successfully return NULL, and for that to safely get returned to
vboot. We need to switch back to regular malloc here instead of using xmalloc.

BUG=None
TEST=Did a recovery boot on nyan and saw that malloc no longer fails.
BRANCH=None

Change-Id: I9d67879252656aa1a0f12652da8a5c8237f605d1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/178708
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>