serial: Expose some functions which are prototyped in libpayload.h.
Some of the serial functions in the x86/8250 serial driver are prototyped in
libpayload, but they weren't exposed in the two ARM serial drivers we have
currently. Since being able to send and receive characters directly through
the serial console is useful for a GDB stub, and because it makes things more
consistent between the drivers in libpayload and in depthcharge, we expose
those functions here as well.

BUG=None
TEST=Built and booted on nyan, built and booted on link, built for pit.
BRANCH=None

Change-Id: I37dd627970b5852dc2214a9d570b5d9a8633be48
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/179604
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>