rambi: Add device NVS layout and support eMMC in ACPI mode
There is a structure in ACPI NVS region that contains the
base addresses for devices that can be put in ACPI mode and
will disappear from the PCI bus.

This structure is directly imported from coreboot and used
here to get at the same set of data that ACPI will present
to the OS, without adding a more messy ACPI parser.  Since
coreboot uses #defines I did not change it to use enum as
it is useful to directly compare it to the coreboot source.

If the device is enabled in ACPI mode then it will be added
directly from the memory address saved in BAR0, otherwise it
will be probed as a standard PCI device.

BUG=chrome-os-partner:24380
BRANCH=none
TEST=build and boot depthcharge on rambi with the eMMC in
both PCI and ACPI modes.

Change-Id: I143ae238c670e391036386f6e58991acafb0ea76
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179884
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>