Build depthcharge will all possible debug information.
Turn on the -ggdb option which builds with "the most expressive format
available", and append a 3 to turn the debugging output up to 3, currently the
highest level supported. This debugging information is stripped out by
cbfstool when depthcharge is installed in an image as a payload, so there's no
harm in piling it on for use on the host.

BUG=None
TEST=Built and booted on nyan.
BRANCH=None

Change-Id: I4bbb8cf3617f7a510725c6c4f66310cccfe065e2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/180176
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>