sdhci: Advertise controller HC capability to the devices
Some eMMC devices (Hynix 32G eMMC in particular) do not operate
properly if the controller does not advertise the high capacity (HC)
capability. The mmc code advertises this capability based on a bit
setting in the host structure, but the SDHCI controller initialization
code was not setting this bit.

One can presume that all controllers used by Chrome OS devices are
going to be HC capable, so the fix is to advertise this capability
unconditionally.

BUG=chrome-os-partner:24675
TEST=manual

   . verified that eMMC boots smoothly with Hynix and Samsung eMMC
     devices.

Change-Id: I0aae35fa44a2fecd52a38a8b0e729d769f6afaf6
Original-Change-Id: Ie802c683f1325ef68830e458aa1462e4370a0d2e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183982
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>