asix: Fix RxCtrl bitfield definitions
This patch fixes the definitions for the RxCtrl bitfield of ASIX network
dongles, which were incorrect for the Accept Broadcast (AB) flag. This
caused the dongle to not relay packets addressed to the Ethernet
broadcast address (ff:ff:ff:ff:ff:ff), and caused netboot to fail DHCP
with servers that sent their Offer to this address. Most servers
directly address Offers to the MAC address they received a Discover
from, which is how this bug could go unnoticed for so long.

BUG=None
TEST=Set up a dnsmasq DHCP server with the --dhcp-broadcast flag,
confirm that netboot can still get a DHCP address from it.

Change-Id: Id3497622b1b374576b0f62d4e965b25cf4d0f91a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182523
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>