flags: Use "die" to handle flag errors instead of returning an error code.
If there are errors setting up flags, that's a programming error which must be
corrected. There's no reason to return an error, because that's not the sort
of problem the calling code will be able to or should try to recover from.

BUG=None
TEST=Built and booted on nyna. Built for all other affected boards.
BRANCH=None

Change-Id: Idedcf9f55f2c60c1634bf22f5bb6c9e2187336c8
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/183172
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>