ARM: Prepare for memsetting addresses above 4GB.
Historically we haven't needed to zero unused memory above 4GB because the ARM
machines we were targeting didn't have enough memory to extend into that
address range. That isn't going to be true forever, so this change rearranges
the ARM arch_phys_memset function so that it will memset addresses which can be
accessed normally using the normal implementation of memset, and has a place
to put a different implementation which will memset the rest using some other
mechanism.

Until the high addresses are handled properly, depthcharge will print a
warning if it was asked to clear something it isn't able to.

BUG=None
TEST=Built and booted on nyan-rev1 in normal, developer and recovery mode and
saw that in developer and recovery modes memory was cleared. Added printfs to
show what ranges were being cleared and saw that the ranges were correct.
Added the following code to wipe_unused_memory in src/vboot/util/memory.c with
the call to memset commented out and verified that the right bit of code tried
to memset the right region of memory.

	printf("Test 0xffffffff, 1\n");
	arch_phys_memset(0xffffffff, 0, 1);
	printf("Test 0xffffffff, 2\n");
	arch_phys_memset(0xffffffff, 0, 2);
	printf("Test 0xffffffffffffffff, 2\n");
	arch_phys_memset(0xffffffffffffffffULL, 0, 2);

Change-Id: Iff1e78e304a98c2e507b6e337a175be9ee5ae193
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/183555
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>