sound: Tolerate sound_ops not being set
If audio is not yet working on a board then it will not set up
sound_ops and the sound driver should tolerate this rather than
calling die() and refusing to boot.

This still does assert if sound_ops is set but the actual ops
are invalid.

BUG=chrome-os-partner:25250
BRANCH=None
TEST=build and boot on rambi in developer mode and don't hang
when it tries to call into any of the sound functions.

Change-Id: I7fad035195ae1681a228503759afac78b6fa18c0
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183673
Reviewed-by: Aaron Durbin <adurbin@chromium.org>