nyan: big: Move flash regions to free 1MB for legacy mode and 16K for elog.
We intend to add support for eventlog on nyan, and to make it possible to
install payloads which can be run using the legacy boot mode (Ctrl+L). The
size of the firmware regions were shrunk to make room for the legacy region,
some unused space was repurposed for the event log, and the few small RW
sections which aren't part of the A or B images were distributed to be after
both the A and B images. This leaves those two larger sections aligned on a
nice boundary and the same size, but gets rid of the empty space after image
A.

The size of the RW firmware should still be considerably smaller than the
space allocated to it by several times. The space for the EC image, the
firmware IDs, and the vblocks were left the same.

BUG=None
TEST=Built and booted in normal mode on nyan.
BRANCH=None

Change-Id: I5734c57d18f59511908c0ea2aeb72416d618ba80
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/183737
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>