Add console history capability
Add a circular array to keep the console command history. When 'arrow
up/down' keys are pressed, cycle through the array looking for the
earlier/later history items.

Allow the user to edit history lines saving cursor position in the
edited line, but do not allow to alter history - editing is good only
if the user wants to run the modified command.

If compiled with HISTORY_DEBUG defined - add a CLI command to help
debugging console history functionality.

BUG=chromium:338061
TEST=manual

  . tried running console on rambi adding numerous commands to history
    and examining history.

Change-Id: I08907cf2e6111179da07dc4f42a77e531089fc47
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184549
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>