arm: Fix accesses to bogus SVC stack pointer
This patch fixes a bug that has been long lurking in the ARM Linux
handoff code, but was only triggered by some recent (seemingly
innocuous) reordering. Since we execute all of firmware (from the start
of coreboot on) in System mode to simplify exception handling, we never
set up the stack pointer (which is individually banked for all modes) in
SVC mode. Linux doesn't expect the stack pointer to be initialized on
entry, so we are fine there... but if we switch modes in C we risk the
trailing end of depthcharge (including possible compiler reorderings) to
acccess the stack after the mode switch, causing data aborts or worse
depending on what value was left in there. Let's move this critical code
to assembly so that we have full control about the order in which things
are executed.

BUG=None
TEST=Booted on Dave's formerly failing Nyan successfully.

Change-Id: Ic49e2fc37e7fa6eb188d8d73909c75fd8280a7e2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184708
Reviewed-by: Gabe Black <gabeblack@chromium.org>