i2c: designware: Allow RESTART conditions to be sent
Set the CONTROL_RE / IC_RESTART_EN bit of our control register to allow
RESTART conditions (including sending a start byte!) to be sent.
According to the datasheet, this may break support with some older slave
devices, but is necessary to support our audio codec (and any I2C slave
that we communicate with by first sending a start byte). This seems like
it should be the standard default, and we shouldn't need to add a
configuration option for it unless we ever need to communicate with such
a legacy device.

BUG=chrome-os-partner:24309
BRANCH=Rambi
TEST=Manual on Rambi.
- Setup designware bus at address 0xd0820000
- Add max98098 codec to bus
- Call into codec driver to perform init
- Read back max98098 part revision over i2c, and verify 'D' is correctly
  read (previously read 0).

Change-Id: I5aa89023d7f82315fcad6558855363060993fbce
Original-Change-Id: I1e3820e00e8cac4169b53cef2e122e485c046a82
Reviewed-on: https://chromium-review.googlesource.com/185896
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>