Improve debug facilities and error handling of the mmc driver.
This is a mostly cosmetic change which makes the following
modifications:

Reduce the noise added by the mmc driver when tracing is enabled.
Print the mmc structure address when tracing the commands, this allows
to associate commands with devices. Print the product description as
reported by devices through the CID register.

Also, prevent memory leak in mmc_setup_media();

BUG=chrome-os-partner:24396
TEST=manual
    . ran the device with debug printouts enabled, see less noisy output

    . with later patches, which fix SD card operation, see proper
      product IDs:

   rambi: sto init
   Man 000045 Snr 1132044051 Product SEM16G Revision 0.7
   Man 000003 Snr 3266950656 Product SU16 Revision 0.2
   *  0: SDHCI fixed
      1: SDHCI card
   2 devices total

Change-Id: I649092ec81baaa381230b508e970a0d219d5cd5f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185794
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>