arm: RW: Fix the RW trampoline on ARM.
We recently started telling the ARM version of libpayload where the coreboot
tables were by passing a pointer to the in %r0. Unfortunately we didn't set
%r0 ourselves when starting the RW version of depthcharge, so libpayload
wouldn't know where to find the tables and may have silently done something
bad which caused it to crash.

Because the load_elf function is shared between x86 and ARM and is where
control actually passes to the RW firmware, we have to add a parameter to it
which is only meaningful (currently) on ARM and which is harmless ignored on
x86. This is a bit of a hack, but since we're moving away from starting the RW
firmware from the RO depthcharge all this code will go away at some point in
the relatively near future.

BUG=chrome-os-partner:26060
TEST=Booted into RW depthcharge on nyan which was previously not working.
BRANCH=None

Change-Id: Ifa57fc7b98be21eb05d1c73b072bbd694938a9ce
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/187383
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>