Add Squawks and Glimmer SD card support
This is the fb9e612906dc patch pasted into the two new boards (the
original change was for Rambi). Below is the original description:

Rambi hardware has an SD card socket connected to the appropriate
Storage Control Cluster port.

Add initialization code to register the SD card port to be able to
boot from the SD card in recovery mode.

BUG=chrome-os-partner:24396
TEST=manual
   . this change is tested with the rest of the patches in the stack.
     Rambi successfully boots in recovery mode from the SD card.

Change-Id: Ifb8ee8cc78ad3292b69d472e54b54042866f478d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187939
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>