netboot: Add option to change kernel command line file name
This patch expands the feature to read the kernel command line over TFTP
by providing a new netboot parameter to specify the filename that the
command line is read from. It slightly changes the precendence order to
determine where the command line comes from: If the new parameter is
provided, it will always be used (and the old KernelArgs parameter is
ignored). Otherwise, depthcharge will fall back to the built-in default
command line (meant for the factory), optionally appending the contents
of KernelArgs if it exists.

BUG=None
TEST=In combination with other two patches, get an image that will boot
off Andrew's TFTP server straight out of chromeos-bootimage.

Change-Id: I9f11b8d7855d7ffbffd902f61ac7d177beb418d6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/197539
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>