zako: Support checking AC adapter type and alert if adapter is wrong.
To detect and alert user that an incompatible or insufficient AC adapter is
connected, we have to render the firmware bitmap screen and then shutdown.

The detection is done by checking a specific GPIO (48) which is reserved for
reporting AC adapter type.

BUG=chrome-os-partner:25024
TEST=emerge-zako depthcharge chromeos-bootimage
BRANCH=zako

Change-Id: I99e3449ea307bb740e3369ce7f5581799c11eace
Reviewed-on: https://chromium-review.googlesource.com/190684
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>