nyans: adjust fmap.dts for RW ramstage verification
In order to vboot verify ramstage there needs to be
something to verify. Add the ramstage as a component
to the RW firmware sections. In addition remove the RO
normal flag from the preamable flags as that is no longer
applicable.

BUG=chrome-os-partner:27094
BRANCH=None
CQ-DEPEND=CL:191535
CQ-DEPEND=CL:191622
TEST=Built nyan. Verified resulting image.bin contained the
     correct number of components.

Change-Id: I01400026a33240358df7f4055bffb206a876ac89
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191031
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>