i2c: exynos: Add support for the "transfer" callback to the controller drivers.
The i2c API is transitioning away from the read and write functions to a more
generic transfer function which will allow structuring transactions more in
line with what slave devices actually expect. This change implements that new
function for the two exynos I2C controller drivers.

The exynos5_usi driver is actually not compliant with the API in that I think
it sends a stop bit after each segment instead of only the last one. To get it
to behave properly we might have to rewrite it to use manual mode instead of
auto mode. Since the behavior of the driver will still be the same for
existing consumers and because active development isn't happening on devices
which use this driver, it should be safe to leave it that way for now and
revisit it later as time allows or circumstances require.

BUG=chrome-os-partner:27097
TEST=Built and booted on daisy and peach_pit with changes to the slave devices
so that they use the new callback instead of the old ones.
BRANCH=None

Change-Id: Idf4f257b827421a04fbfe305eca136289d2eb6d1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/191636
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>