tpm: i2c: Convert the driver over to the new "transfer" based interface.
The driver calls into the new interface primarily through convenience
wrappers, but also directly in the case of reads on the SLB9645. The pattern
there is similar to the i2c_readb wrapper, except that it needs to read an
array of data instead of a single byte.

BUG=chrome-os-partner:27097
TEST=Built and booted on nyan. Verified that TPM communication worked.
BRANCH=None

Change-Id: Ia8b5824634df8c8f1dcb8f8f4f896cec1ea4bc34
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/191661
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>