peach_pit: Don't include the ramstage in the RW sections.
The unified_depthcharge USE flag isn't turned on for peach_pit, so the
cros_bundle_firmware isn't told where to get the ramstage blob. Since we're
not going to use the ramstage and it breaks chromeos-bootimage when coreboot
is used, lets take it back out.

BUG=None
TEST=Built and booted on peach_pit with a proposed cros_bundle_firmware
change. Before this chromeos-bootimage failed, and after it worked.
BRANCH=None

Change-Id: I3aa0b472d140bd540f9c44990398822d29665e88
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/191783
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>