spi: Let the bus structure track whether we need to initialize the controller.
The initialization function was maintaining its own static variable which
gated whether to initialize the controller. The SPI controller structure also
maintains a variable for the same purpose, and we want to keep track of this
on a per controller basis. Otherwise only the first controller will be
initialized.

BUG=None
TEST=Built and booted on daisy and peach_pit.
BRANCH=None

Change-Id: I461fb759d55f6ab9674b5aba9369523852c1b6ce
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/191784
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>