Add some code to set up a ramoops area on device tree based systems.
This code removes any single existing ramoops node in the device tree, adds
a new one with the given parameters, adds the ramoops buffer area to the list
of reserved regions in the device tree, and tells the memory wiping code not
to wipe it.

Like with drivers, this code is enabled using a kconfig option and then
initialized by a function called with the appropriate parameters while
depthcharge is starting.

BUG=chrome-os-partner:27537
TEST=Built and booted on nyan with this code enabled in normal and developer
modes (with and without memory wiping) and verified that when the kernel
paniced, its log output was preserved by the ramoops mechanism and could be
retrieved when the system rebooted.
BRANCH=None

Change-Id: I4ff323e0d9e7809d68820fde3792d07c76add2e3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/193995
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>