ipq806x: add SPI driver
This is a port of the coreboot implementation of the ipq806x SPI
controller driver. It was modified to plug in into depthcharge driver
model and not to use integer addresses in io accessor functions.

For some reason depthcharge does not expect the SPI driver to be
capable of reading after write in a single CS activation, so the
ported driver needed to be modified to keep the CS asserted across
multiple spi_xfer() invocations.

BUG=chrome-os-partner:27784

TEST=was able to read the SPI flash contents using the later
    introduced sf CLI command:

  storm: sf dump 0 10

  00000000: d1 dc 4b 84 34 10 d7 73 15 00 00 00 ff ff ff ff
  storm: sf dump 220000 10

  00220000: 05 00 00 00 03 00 00 00 00 00 00 00 00 00 60 40
  storm:

Change-Id: I61057253c0ce4c1b76a00535cee1268e594f5990
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202340
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>